(#273) Move reminder commands to src/cmd/reminder.py
